Add grid size option
Add option to select if you want the grid to be small, dynamic (what it is today) or large.

Small is 4 column design and large is 2 column design.

This option is global for the module so if selecting small or large the size option on each card gets hidden.

The option is default hidden so themes that support it should enable it.